WEEK-2 REGRESSION
In-Class Exercise:- 
Recall from c02_grad_desc_models.ipynb that the polynomial model fits poorly to the test data - even after attempting L1, L2, and ElasticNet regularization. For this exercise, add to the notebook an "early stop" approach to fitting the polynomial (and concatenate this result to the rmses data frame). Finally, add a markdown cell at the end of the notebook discussing the results of the early stop approach for this model. 

Weekly Exercise:- 
Choose one of the models from 1 through 4 (c02_model1, c02_model2, etc.).
Create two copies of your chosen model-generating notebook and the model-fitting notebook. 
First copy- Increase the amount of noise added to the data in your data-generating notebook. Run the model fitting notebook on this new data. Add a markdown cell at the end of the notebook that discusses how the resulting model fits results and how these compare to the original.
Second copy- Decrease the number of observations generated in the data-generating notebook. Run the model fitting notebook on this new data. Add a markdown cell at the end of the notebook that discusses how this resulting model fits results and how these compare to the original.