WEEK-3 SUPPORT VECTOR MACHINES
In-class Exercise:- 
Start with a copy of the 5_airbnb_logistic_model.ipynb. Add to your analysis the SVM models used in the SVM example (svm with linear kernel, svm with rbv kernel, and svm with poly kernel). Add these models' Accuracy, Precision, Recall, and F1 results to your performance data frame. Summarize the results at the end.

Weekly Exercise:-
First: Develop a predictive model that predicts lawn mower ownership (use the RisingMower.csv data) using SVM classification and the three kernels demonstrated in class today. Display the results of each of these models (accuracy, precision, recall, and F1) and save the 'winning' model to a pickle file.
Second: Create an application that asks the user for an income and lot size and answers if it's predicted that this property would own a lawnmower.