WEEK-6 TEXT MINING
This week I learned how to deal with the data imbalance in the data and also how to solve the Mult class problems and text mining and I have practiced them.

In-Class Exercise:-
Update the Text Mining tutorial by adding the lemmatization technique in Text Mining Fundamentals. Recall that lemmatization of the data will result in counting begin, beginning, begun, began as 'begin'. This will reduce the overall size of the dataset, and focus on the root of the given word. 
NOTE: This should be done during the preprocessing stage; before the data split.

Weekly Exercise:-
In the weekly exercise, you will explore the impacts of applying SVD to the data.  Analyze how changing the n_components parameter impacts the modeling results. Use the values 100, 300, and 500 and discuss how each of these values impacted the performance of the models. Discuss why we may or may not want to use SVD in our analysis.