WEEK-9 CONVOLUTIONAL NEURAL NETWORKS
This week I have learned about Convolutional Neural Networks and practiced them. There are a lot of files for the images and log which I could not upload in this folder as it's too large. The following contains all the files for CNN-"https://usf.box.com/s/7bn9cox3sqyyw32lqqps2cb4nmcq0buv".
There is no In Class or Weekly Exercises this week.